CRUD im Editor

+++ app/client/src/main/svelte/pages/OwnerEditor.svelte
@@ -1,5 +1,9 @@
 <script>
     import { createEventDispatcher } from 'svelte';
+	import { toast } from '../components/Toast';
+	import { createValue } from '../utils/rest.js';
+	import { updatePatch } from '../utils/rest.js';
+	import { removeValue } from '../utils/rest.js';
 	import Button from '../components/Button';
 	import TextField from '../components/TextField';
     
@@ -9,14 +13,13 @@
     let showUpdate;
     let showRemove;
     let newOwner = {
-        name: undefined,
-        address: undefined,
-        contact: undefined
+        name: '',
+        address: '',
+        contact: ''
     }
 
-    $: disabled = !newOwner.name || !newOwner.address || !newOwner.contact;
-    $: if (owner) onChange()
-    function onChange() {
+    $: if (owner) onChangeOwner()
+    function onChangeOwner() {
         showUpdate = true;
         showRemove = true;
         newOwner = {
@@ -25,24 +28,50 @@
             address: owner.address,
             contact: owner.contact
         }
-        console.log(['onChange', newOwner]);
+        console.log(['onChangeOwner', newOwner]);
     }
 
+    $: disabled = !newOwner.name || !newOwner.address || !newOwner.contact;
+
     const dispatch = createEventDispatcher();
-    function onCreate() {
-        visible = false;
-        console.log(['create', newOwner]);
-        dispatch('create', newOwner);
+    function onCreateOwner() {
+        createValue('/api/owner', newOwner)
+        .then(json => {
+            console.log(['onCreateOwner', newOwner, json]);
+            visible = false;
+            dispatch('create', newOwner);
+        })
+        .catch(err => {
+            console.log(['onCreateOwner', newOwner, err]);
+            toast.push(err.toString());
+        });;
     }
-    function onUpdate() {
-        visible = false;
-        console.log(['update', newOwner]);
-        dispatch('update', newOwner);
+    function onUpdateOwner() {
+        updatePatch('/api/owner' + '/' + newOwner.id, newOwner)
+        .then(json => {
+            console.log(['onUpdateOwner', newOwner, json]);
+            visible = false;
+            dispatch('update', newOwner);
+        })
+        .catch(err => {
+            console.log(['onUpdateOwner', newOwner, err]);
+            toast.push(err.toString());
+        });;
     }
-    function onRemove() {
-        visible = false;
-        console.log(['remove', newOwner]);
-        dispatch('remove', newOwner);
+    function onRemoveOwner() {
+        const text = newOwner.name;
+        const hint = text.length > 20 ? text.substring(0, 20) + '...' : text;
+		if (!confirm("Really delete '" + hint + "'?")) return;
+        removeValue('/api/owner' + '/' + newOwner.id)
+        .then(() => {
+            console.log(['onRemoveOwner', newOwner]);
+            visible = false;
+            dispatch('remove', newOwner);
+        })
+        .catch(err => {
+            console.log(['onRemoveOwner', newOwner, err]);
+            toast.push(err.toString());
+        });;
     }
     function onCancel() {
         visible = false;
@@ -51,17 +80,20 @@
 
 <div class="flex flex-col">
     <div class="w-full">
-        <TextField bind:value={newOwner.name} 
+        <TextField 
+            bind:value={newOwner.name} 
             label="Name"		
             placeholder="Insert a name"/>
     </div>
     <div class="w-full">
-        <TextField bind:value={newOwner.address} 
+        <TextField 
+            bind:value={newOwner.address} 
             label="Address"		
             placeholder="Insert a text"/>
     </div>
     <div class="w-full">
-        <TextField bind:value={newOwner.contact} 
+        <TextField 
+            bind:value={newOwner.contact} 
             label="Contact"		
             placeholder="Insert a text"/>
     </div>
@@ -69,16 +101,16 @@
 
 <div class="py-4">
     {#if showUpdate}
-    <Button on:click={() => onUpdate()} {disabled}>
+    <Button on:click={() => onUpdateOwner()} {disabled}>
         Ok
     </Button>
     {:else}
-    <Button on:click={() => onCreate()} {disabled}>
+    <Button on:click={() => onCreateOwner()} {disabled}>
         Ok
     </Button>
     {/if}
     {#if showRemove}
-    <Button on:click={() => onRemove()}>
+    <Button on:click={() => onRemoveOwner()}>
         Löschen
     </Button>
     {/if}
